refactor(core): move review platform owner to services
