+++ src/crates/services/services-integrations/src/lib.rs
@@ -40,7 +40,10 @@ pub mod remote_connect;
 pub mod remote_ssh;
 
 #[cfg(feature = "review-platform")]
-pub mod review_platform_http;
+pub mod review_platform;
+
+#[cfg(feature = "review-platform")]
+pub(crate) mod review_platform_http;
 
 #[cfg(feature = "workspace-search")]
 pub mod workspace_search;
